Add ODR fitter, and shortest d to line func.
This commit includes a new fitter for linear lines, such as atom planes.
The ODR-fitter is better than the old OLS fitter for vertical lines.
Also, a new function to find the shortest distance from a point to a
line given by a slope and intercept. Returs the distance with sign
(+/-).