Change how get_closest_distance_and_angle_to_point works
The function get_closest_distance_and_angle_to_point for getting a list
of shortest distances from a list of atom positions to an atom plane is
changed: The function now uses get_shortest_distance_point_to_line from
atomap.fitting_tools to get the shortest distance. This function uses
an analytical expression of the line, and finds the distance for all
atom positions by array operations (looping through the position list).
Negative distances are allowed, making this function much simpler!
The generation of the debug plot is also changed: plots a 3D scatter
plot of position and distance to plane. For the plane the analytical
expression is plotted as a line evaluated for 1000 points, and the atom
position of the atoms in the plane are also plotted as scatter points.