Atom layer clustering: Docstring, bugfix, expansion
Bugfix: last layer in combine_clustered_positions_into_layers
was never appended. Is appended now. find_average_distance_between_atoms
is expanded to also find the separation of all monolayers, and the
average, in addition to the first. Still the first is used in later
processing. It could be possible to assess if the planes were well
indentified (though monolayer separation std) by knowing more about
the variance in atom position. Some more functionality is needed in
MakeTestData to test this systematically. A warning is now issued if
there is an obvious outlier corrupting the data, but the limit is not
fine tuned.