unit tests: make it possible to test Qt5Agg GUI elements in matplotlib
Main issue was making the virtual display xvfb-run work ith Qt5Agg.
The default setting in xvfb-run did not work, so --server-args
had to used to set a supported display setting.

Currently this will only work with the matplotlib qt backend, and
not the nbagg matplotlib notebook backend, since the backend has to
be set before importing matplotlib.pyplot. Thus the method used in
this commit is not ideal, but will suffice for now.