gitlab-ci.yml: split pages into release and master jobs
Currently not possible to place the output from the pages job in any
other place than a folder called "public". This means the webpage for
the master (development branch) cannot be deployed in the pages job,
since putting it in the "public" folder would overwrite the release
webpage.

In the process of being implemented:
https://gitlab.com/gitlab-org/gitlab/issues/16208